+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Fix typo in redshift function
 - Update to sklearn v1.1.3
+- add an EoS marginalization option to source_classification_pe. Draws
+  are from Legred et al. (https://zenodo.org/record/6502467#.Y1xFdHbMI2z)
 
 ## [1.0.4] - 2022-10-19